Add files via upload
Practice:
Write  A program consisting of four parts and a menu unifying them. In the main menu, the program should display the following question to the user:
Choose an option
The user will choose a number between 1 and 3.
1: Hourglass
2: All numbers greater than 10
3: Number count
4: Exit
Any other number: Print 'Input error' to the screen and return to the main menu (Choose an option).
As long as the user has not chosen option number 4, the main menu question should be displayed to the user.
(Assume the user will input an integer and nothing else)
a. Hourglass:
The program will print to the screen:
Please enter number of rows
and will receive from the user a positive integer (assume the input is valid) representing the height of the hourglass.
The program should print an hourglass to the screen using asterisks.
Note that the number of asterisks in the first and last rows of the hourglass should be equal.
For even and odd inputs, the hourglass will be drawn differently.
Important note: Do not use the multiplication operation to replicate a string multiple times (i.e., it is forbidden to perform the operation: print('*' * 7))

b. All numbers greater than 10:
The program will print to the screen in a loop:
Please enter a number
and will receive an integer from the user each time (assume the input is valid), until the user inputs -1.
After the user inputs -1, the program should print in one column (each number on a new line) all numbers greater than 10 in ascending order along with the message: The numbers are

c. Number count:
The program will print to the screen in a loop:
Please enter a number
and will receive an integer from the user (assume the input is valid).
After receiving the number, the program will again prompt the user to enter a number, until the user inputs -1.
After the user inputs -1, the program should print in one column (each number on a new line) all numbers the user entered along with the number of times each number was entered, along with the message: The numbers are
Note: The numbers should be printed in the order they were entered.

d. Exit:
The program should print 'Bye' to the screen and terminate.